chore: rename index to core
